@@ -10,7 +10,7 @@
         }
     ],
     "require": {
-        "php": "^7.4",
+        "php": "^8.1",
         "symfony/framework-bundle": "^5.4",
         "spipu/ui-bundle": "^2.0"
     },
