@@ -15,8 +15,7 @@ namespace Spipu\UserBundle\Controller;
 
 use Doctrine\ORM\EntityManagerInterface;
 use Exception;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
-use Spipu\UiBundle\Exception\UiException;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Spipu\UiBundle\Service\Ui\FormFactory;
 use Spipu\UiBundle\Service\Ui\GridFactory;
 use Spipu\UiBundle\Service\Ui\ShowFactory;
@@ -31,43 +30,26 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use Throwable;
 
 /**
  * Class AdminUserController
- * @Route("/user")
  * @SuppressWarnings(PMD.ExcessiveClassComplexity)
  * @SuppressWarnings(PMD.CouplingBetweenObjects)
  * @SuppressWarnings(PMD.TooManyPublicMethods)
  */
+#[Route(path: '/user')]
 class AdminUserController extends AbstractController
 {
-    /**
-     * @var EntityManagerInterface
-     */
-    private $entityManager;
+    private EntityManagerInterface $entityManager;
 
-    /**
-     * @param EntityManagerInterface $entityManager
-     */
     public function __construct(
         EntityManagerInterface $entityManager
     ) {
         $this->entityManager = $entityManager;
     }
 
-    /**
-     * @Route(
-     *     "/",
-     *     name="spipu_user_admin_list",
-     *     methods="GET"
-     * )
-     * @Security("is_granted('ROLE_ADMIN_MANAGE_USER_SHOW')")
-     * @param GridFactory $gridFactory
-     * @param UserGrid $userGrid
-     * @return Response
-     * @throws UiException
-     */
+    #[Route(path: '/', name: 'spipu_user_admin_list', methods: 'GET')]
+    #[IsGranted('ROLE_ADMIN_MANAGE_USER_SHOW')]
     public function index(GridFactory $gridFactory, UserGrid $userGrid): Response
     {
         $manager = $gridFactory->create($userGrid);
@@ -79,20 +61,8 @@ class AdminUserController extends AbstractController
         return $this->render('@SpipuUser/admin/index.html.twig', ['manager' => $manager]);
     }
 
-    /**
-     * @Route(
-     *     "/edit/{id}",
-     *     name="spipu_user_admin_edit",
-     *     methods="GET|POST"
-     * )
-     * @Security("is_granted('ROLE_ADMIN_MANAGE_USER_EDIT')")
-     * @param FormFactory $formFactory
-     * @param UserForm $userForm
-     * @param UserRepository $userRepository
-     * @param int $id
-     * @return Response
-     * @throws UiException
-     */
+    #[Route(path: '/edit/{id}', name: 'spipu_user_admin_edit', methods: 'GET|POST')]
+    #[IsGranted('ROLE_ADMIN_MANAGE_USER_EDIT')]
     public function edit(
         FormFactory $formFactory,
         UserForm $userForm,
@@ -117,21 +87,8 @@ class AdminUserController extends AbstractController
         return $this->render('@SpipuUser/admin/edit.html.twig', ['manager' => $manager]);
     }
 
-    /**
-     * @Route(
-     *     "/create/",
-     *     name="spipu_user_admin_create",
-     *     methods="GET|POST"
-     * )
-     * @Security("is_granted('ROLE_ADMIN_MANAGE_USER_EDIT')")
-     * @param FormFactory $formFactory
-     * @param UserForm $userForm
-     * @param ModuleConfigurationInterface $moduleConfiguration
-     * @param MailManager $mailManager
-     * @return Response
-     * @throws UiException
-     * @throws Throwable
-     */
+    #[Route(path: '/create/', name: 'spipu_user_admin_create', methods: 'GET|POST')]
+    #[IsGranted('ROLE_ADMIN_MANAGE_USER_EDIT')]
     public function create(
         FormFactory $formFactory,
         UserForm $userForm,
@@ -158,21 +115,8 @@ class AdminUserController extends AbstractController
         return $this->render('@SpipuUser/admin/create.html.twig', ['manager' => $manager]);
     }
 
-    /**
-     * @Route(
-     *     "/show/{id}",
-     *     name="spipu_user_admin_show",
-     *     methods="GET"
-     * )
-     * @Security("is_granted('ROLE_ADMIN_MANAGE_USER_SHOW')")
-     * @param ShowFactory $showFactory
-     * @param UserForm $userForm
-     * @param UserRepository $userRepository
-     * @param RoleService $roleService
-     * @param int $id
-     * @return Response
-     * @throws UiException
-     */
+    #[Route(path: '/show/{id}', name: 'spipu_user_admin_show', methods: 'GET')]
+    #[IsGranted('ROLE_ADMIN_MANAGE_USER_SHOW')]
     public function show(
         ShowFactory $showFactory,
         UserForm $userForm,
@@ -199,20 +143,8 @@ class AdminUserController extends AbstractController
         );
     }
 
-    /**
-     * @Route(
-     *     "/update-acl/{id}",
-     *     name="spipu_user_admin_acl",
-     *     methods="POST"
-     * )
-     * @Security("is_granted('ROLE_ADMIN_MANAGE_USER_SHOW')")
-     * @param UserRepository $userRepository
-     * @param Request $request
-     * @param EntityManagerInterface $entityManager
-     * @param RoleService $roleService
-     * @param int $id
-     * @return Response
-     */
+    #[Route(path: '/update-acl/{id}', name: 'spipu_user_admin_acl', methods: 'POST')]
+    #[IsGranted('ROLE_ADMIN_MANAGE_USER_SHOW')]
     public function updateAcl(
         UserRepository $userRepository,
         Request $request,
@@ -245,18 +177,8 @@ class AdminUserController extends AbstractController
         return $redirectResponse;
     }
 
-    /**
-     * @Route(
-     *     "/delete/{id}",
-     *     name="spipu_user_admin_delete",
-     *     methods="DELETE"
-     * )
-     * @Security("is_granted('ROLE_ADMIN_MANAGE_USER_DELETE')")
-     * @param Request $request
-     * @param UserRepository $userRepository
-     * @param int $id
-     * @return Response
-     */
+    #[Route(path: '/delete/{id}', name: 'spipu_user_admin_delete', methods: 'DELETE')]
+    #[IsGranted('ROLE_ADMIN_MANAGE_USER_DELETE')]
     public function delete(
         Request $request,
         UserRepository $userRepository,
@@ -294,18 +216,8 @@ class AdminUserController extends AbstractController
         return $this->redirectTo('list');
     }
 
-    /**
-     * @Route(
-     *     "/enable/{id}/{backTo}",
-     *     name="spipu_user_admin_enable",
-     *     methods="GET"
-     * )
-     * @Security("is_granted('ROLE_ADMIN_MANAGE_USER_EDIT')")
-     * @param UserRepository $userRepository
-     * @param int $id
-     * @param string $backTo
-     * @return Response
-     */
+    #[Route(path: '/enable/{id}/{backTo}', name: 'spipu_user_admin_enable', methods: 'GET')]
+    #[IsGranted('ROLE_ADMIN_MANAGE_USER_EDIT')]
     public function enable(
         UserRepository $userRepository,
         int $id,
@@ -338,18 +250,8 @@ class AdminUserController extends AbstractController
         return $this->redirectTo($backTo, $resource);
     }
 
-    /**
-     * @Route(
-     *     "/disable/{id}/{backTo}",
-     *     name="spipu_user_admin_disable",
-     *     methods="GET"
-     * )
-     * @Security("is_granted('ROLE_ADMIN_MANAGE_USER_EDIT')")
-     * @param UserRepository $userRepository
-     * @param int $id
-     * @param string $backTo
-     * @return Response
-     */
+    #[Route(path: '/disable/{id}/{backTo}', name: 'spipu_user_admin_disable', methods: 'GET')]
+    #[IsGranted('ROLE_ADMIN_MANAGE_USER_EDIT')]
     public function disable(
         UserRepository $userRepository,
         int $id,
@@ -382,19 +284,8 @@ class AdminUserController extends AbstractController
         return $this->redirectTo($backTo, $resource);
     }
 
-    /**
-     * @Route(
-     *     "/reset/{id}",
-     *     name="spipu_user_admin_reset",
-     *     methods="GET"
-     * )
-     * @Security("is_granted('ROLE_ADMIN_MANAGE_USER_EDIT')")
-     * @param UserRepository $userRepository
-     * @param MailManager $mailManager
-     * @param int $id
-     * @return Response
-     * @throws Throwable
-     */
+    #[Route(path: '/reset/{id}', name: 'spipu_user_admin_reset', methods: 'GET')]
+    #[IsGranted('ROLE_ADMIN_MANAGE_USER_EDIT')]
     public function reset(
         UserRepository $userRepository,
         MailManager $mailManager,
@@ -417,46 +308,21 @@ class AdminUserController extends AbstractController
 
         return $this->redirectTo('show', $resource);
     }
-    /**
-     * @Route(
-     *     "/mass-enable",
-     *     name="spipu_user_admin_mass_enable",
-     *     methods="POST"
-     * )
-     * @Security("is_granted('ROLE_ADMIN_MANAGE_USER_EDIT')")
-     * @param UserRepository $userRepository
-     * @param Request $request
-     * @return Response
-     */
+
+    #[Route(path: '/mass-enable', name: 'spipu_user_admin_mass_enable', methods: 'POST')]
+    #[IsGranted('ROLE_ADMIN_MANAGE_USER_EDIT')]
     public function massEnable(UserRepository $userRepository, Request $request): Response
     {
         return $this->massAction($userRepository, $request, 'enable', 'spipu.user.success.mass_enabled');
     }
 
-    /**
-     * @Route(
-     *     "/mass-disable",
-     *     name="spipu_user_admin_mass_disable",
-     *     methods="POST"
-     * )
-     * @Security("is_granted('ROLE_ADMIN_MANAGE_USER_EDIT')")
-     * @param UserRepository $userRepository
-     * @param Request $request
-     * @return Response
-     */
+    #[Route(path: '/mass-disable', name: 'spipu_user_admin_mass_disable', methods: 'POST')]
+    #[IsGranted('ROLE_ADMIN_MANAGE_USER_EDIT')]
     public function massDisable(UserRepository $userRepository, Request $request): Response
     {
         return $this->massAction($userRepository, $request, 'disable', 'spipu.user.success.mass_disabled');
     }
 
-    /**
-     * @param UserRepository $userRepository
-     * @param Request $request
-     * @param string $action
-     * @param string $transLabel
-     * @return Response
-     * @SuppressWarnings(PMD.CyclomaticComplexity)
-     */
     private function massAction(
         UserRepository $userRepository,
         Request $request,
@@ -488,11 +354,6 @@ class AdminUserController extends AbstractController
         return $this->redirectTo('list');
     }
 
-    /**
-     * @param UserInterface $row
-     * @param string $action
-     * @return bool
-     */
     private function massActionRow(UserInterface $row, string $action): bool
     {
         if ($this->getUser()->getId() === $row->getId()) {
@@ -513,11 +374,6 @@ class AdminUserController extends AbstractController
         return false;
     }
 
-    /**
-     * @param string $backTo
-     * @param UserInterface|null $resource
-     * @return Response
-     */
     private function redirectTo(string $backTo, UserInterface $resource = null): Response
     {
         switch ($backTo) {
@@ -533,31 +389,17 @@ class AdminUserController extends AbstractController
         }
     }
 
-    /**
-     * @param string $type
-     * @param string $message
-     * @param array $params
-     * @return void
-     */
     private function addFlashTrans(string $type, string $message, array $params = []): void
     {
         $this->addFlash($type, $this->trans($message, $params));
     }
 
-    /**
-     * @param string $message
-     * @param array $params
-     * @return string
-     */
     private function trans(string $message, array $params = []): string
     {
         return $this->container->get('translator')->trans($message, $params);
     }
 
-    /**
-     * @return array
-     */
-    public static function getSubscribedServices()
+    public static function getSubscribedServices(): array
     {
         return parent::getSubscribedServices() + [
             'translator',
