@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Spipu\UserBundle\Entity;
 
-use DateTime;
+use DateTimeInterface;
 use Serializable;
 use Spipu\UiBundle\Entity\EntityInterface;
 use Spipu\UiBundle\Entity\TimestampableInterface;
@@ -27,23 +27,10 @@ interface UserInterface extends
     Serializable,
     PasswordAuthenticatedUserInterface
 {
-    /**
-     * Get the PK id
-     * @return int|null
-     */
     public function getId(): ?int;
 
-    /**
-     * Get the email
-     * @return null|string
-     */
     public function getEmail(): ?string;
 
-    /**
-     * Set the email
-     * @param string $email
-     * @return self
-     */
     public function setEmail(string $email): self;
 
     /**
@@ -58,116 +45,43 @@ interface UserInterface extends
      */
     public function getUsername();
 
-    /**
-     * @return string|null
-     */
     public function getUserIdentifier(): ?string;
 
-    /**
-     * Set the username
-     * @param string $username
-     * @return self
-     */
     public function setUsername(string $username): self;
 
-    /**
-     * Set the password
-     * @param string $password
-     * @return self
-     */
     public function setPassword(string $password): self;
 
-    /**
-     * Get the PlainPassword
-     * @return string|null
-     */
     public function getPlainPassword(): ?string;
 
-    /**
-     * Set the PlainPassword
-     * @param string $password
-     * @return self
-     */
     public function setPlainPassword(string $password): self;
 
-    /**
-     * Get the FirstName
-     * @return null|string
-     */
     public function getFirstName(): ?string;
 
-    /**
-     * Set the FirstName
-     * @param string $firstName
-     * @return self
-     */
     public function setFirstName(string $firstName): self;
 
-    /**
-     * Get the LastName
-     * @return null|string
-     */
     public function getLastName(): ?string;
 
-    /**
-     * Set the LastName
-     * @param string $lastName
-     * @return self
-     */
     public function setLastName(string $lastName): self;
 
     /**
-     * Set the roles
      * @param string[] $roles
      * @return self
      */
     public function setRoles(array $roles): self;
 
-    /**
-     * Get the Nb of Login
-     * @return int|null
-     */
     public function getNbLogin(): ?int;
 
-    /**
-     * Set the Nb of Login
-     * @param int $nbLogin
-     * @return self
-     */
     public function setNbLogin(int $nbLogin): self;
 
-    /**
-     * Get the nb of try login with a wrong password
-     * @return int|null
-     */
     public function getNbTryLogin(): ?int;
 
-    /**
-     * Get the nb of try login with a wrong password
-     * @param int $nbTryLogin
-     * @return self
-     */
     public function setNbTryLogin(int $nbTryLogin): self;
 
-    /**
-     * @return bool|null
-     */
     public function getActive(): ?bool;
 
-    /**
-     * @param bool $active
-     * @return self
-     */
     public function setActive(bool $active): self;
 
-    /**
-     * @return DateTime|null ?\DateTime
-     */
-    public function getTokenDate(): ?DateTime;
+    public function getTokenDate(): ?DateTimeInterface;
 
-    /**
-     * @param DateTime|null $tokenDate
-     * @return self
-     */
-    public function setTokenDate(?DateTime $tokenDate): self;
+    public function setTokenDate(?DateTimeInterface $tokenDate): self;
 }
