@@ -17,10 +17,7 @@ use Symfony\Component\Security\Core\Exception\AccountStatusException;
 
 class UnactivatedAccountException extends AccountStatusException
 {
-    /**
-     * @return string
-     */
-    public function getMessageKey()
+    public function getMessageKey(): string
     {
         return 'Unactivated Account';
     }
